Created dedicated LibraryRoot test file and moved rendering tests from  modulestore/tests/test_libraries
